Move all badge related tasks to tasks.py
